// src/app/modules/tecnologia-informacao/estoque/nivel-estoque/models/nivelEstoque.ts
export interface ITecnologiaInformacaoNivelEstoque {
  ID_TEIN_ESTO_NIVL: number
  ID_TEIN_ESTO_PRTP: number;
  ID_TEIN_ESTO_PROD: number;
  ID_TEIN_ESTO_PRMO: number;
  NM_PROD: string;
  NM_EMPR: string;
  DS_UNID_MEDI: string;
  TT_ESTO_MNMO: number;
  TT_ESTO_MXMO: number;
  TT_ESTO_ATUA: number;
  UUID_EMPR: string;
  IN_STAT: string;
  DT_INCL: string;
  DS_OBSE: string;
  IN_PAGI: string;
}