// src/app/modules/tecnologia-informacao/estoque/painel-aprovacao/models/painel-aprovacao.ts
import { NumberSymbol } from '@angular/common';

export interface ITecnologiaInformacaoEstoquePainelAprovacao {
  ID_TEIN_ESTO_PRTP: number;
  ID_TEIN_ESTO_APRO: number;
  ID_TEIN_ESTO_PROD: number;
  SG_SIST_ORIG: string;
  UUID_EMPR: number;
  DS_UNID_MEDI:string;
  NM_TIPO: string;
  ID_EMPR: number;
  DS_APRO: string;
  DT_INCL: string;
  DS_PARE: string;
  IN_PAGI: string;
}